Discord Bot/requirements.txt
discord.py==2.3.2
openai==1.10.0
python-dotenv==1.0.0